allocator/dmabuf: Add `DmabufAllocator` wrapper
